# backend/openapi.yml
openapi: 3.0.0
info:
  title: XConnect API
  description: |
    XConnect - Hyperlocal Print Marketplace API
    
    A platform connecting customers with nearby print shops for document printing services.
  version: 1.0.0
  contact:
    name: XConnect Support
    email: support@xconnect.app
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.xconnect.app/api
    description: Production server
...
paths:
...
  /orders/create:
    post:
      tags:
        - Orders
      summary: Create order
      description: Place a new print order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - vendorId
                - document
                - pageCount
                - paymentMethod
                - printConfig
              properties:
                vendorId:
                  type: string
                document:
                  type: string
                  format: binary
                  description: PDF or image file
                pageCount:
                  type: integer
                  minimum: 1
                paymentMethod:
                  type: string
                  enum: [cash, upi]
                printConfig:
                  type: string
                  description: JSON string with paperSize, printType, copies
                useWallet:
                  type: boolean
                  default: false
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Get detailed order information
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        _id:
          type: string
        customerId:
          type: string
        vendorId:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        pageCount:
          type: integer
        printConfig:
          type: object
          properties:
            paperSize:
              type: string
            printType:
              type: string
            copies:
              type: integer
        pricePerPage:
          type: number
        totalAmount:
          type: number
        bookingFee:
          type: number
        walletAmountUsed:
          type: number
        freePagesUsed:
          type: integer
        finalAmount:
          type: number
        pickupToken:
          type: string
        status:
          type: string
          enum: [pending, accepted, printing, completed, rejected]
        payment:
          type: object
          properties:
            method:
              type: string
            status:
              type: string
        createdAt:
          type: string
          format: date-time
...